// src/enums/raw-news-category.enum.ts
export enum RawNewsCategoryEnum {
  POLITICS = "politics",
  WORLD = "world",
  SPORTS = "sports",
  TECHNOLOGY = "technology",
  ECONOMY = "economy",
  ENTERTAINMENT = "entertainment",
  CELEBRITIES = "celebrities",
  HEALTH = "health",
  EDUCATION = "education",
  CULTURE = "culture",
  ENVIRONMENT = "environment",
  SCIENCE = "science",
  SECURITY = "security",
  JUSTICE_AND_LAW = "justice_and_law",
  TRANSPORT = "transport",
  LIFESTYLE = "lifestyle",
  RELIGION = "religion",
  INTERNATIONAL_POLITICS = "international_politics",
}